requirements.txt
beautifulsoup4==4.10.0
numpy==1.15.1
numpy==1.22.0
numpy==1.20.3
pandas==1.3.5
pandas==1.3.4
Pillow==8.4.0
Pillow==10.3.0
Requests==2.32.3
streamlit==1.12.0
